% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{classification_table}
\alias{classification_table}
\title{Binary Logistic Regression: Classification Table}
\usage{
classification_table(model, response)
}
\arguments{
\item{model}{A binary logistic regression model object.}

\item{response}{The dependent variable in \code{model}.}
}
\description{
Binary Logistic Regression: Classification Table
}
\details{
Creates classification table for binary logistic regresison model
  using optimal cut point for accuracy.
}
\examples{
formulas <- create_formula_objects("am", c("hp", "mpg"), c("disp"),
c("drat"))
mtcars_models <- create_model_objects(formulas, data=mtcars,
type="binomial")
last_model <- mtcars_models[[length(mtcars_models)]]
classification_table(last_model, last_model$model[,1])
}

