% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{create_formula_objects}
\alias{create_formula_objects}
\title{Hierarchical Formula Generation}
\usage{
create_formula_objects(outcome, block1, ...)
}
\arguments{
\item{outcome}{The dependent variable of the hierarchical model}

\item{block1}{A character vector, with names of variables. The first block of
independent variables.}

\item{...}{A character vector, with names of variables. Subsequent blocks of
independent variables.}
}
\value{
A list of \code{lm} formulas
}
\description{
Hierarchical Formula Generation
}
\examples{
create_formula_objects("y", c("lag.quarterly.revenue"), c("price.index"))
create_formula_objects("y", c("lag.quarterly.revenue"), c("price.index",
"income.level"))
}

