% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{model_coefficient_table_binomial}
\alias{model_coefficient_table_binomial}
\title{Binary Logistic Regression: Coefficient Table Output}
\usage{
model_coefficient_table_binomial(models)
}
\arguments{
\item{models}{A list of \code{lm} model objects.  A set of model objects
created by \code{create_model_object}.}
}
\description{
Binary Logistic Regression: Coefficient Table Output
}
\details{
Creates table output to summarize model coefficients for all models
  in a hierarchical binary logistic regression analysis.
}
\examples{
formulas <- create_formula_objects("am", c("hp", "mpg"), c("disp"),
c("drat"))
mtcars_models <- create_model_objects(formulas, data=mtcars,
type="binomial")
\dontrun{model_summary_table_binomial(mtcars_models)}
}

