% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{model_output}
\alias{model_output}
\title{Multiple Regression Assumption Checking}
\usage{
model_output(models)
}
\arguments{
\item{models}{A list of \code{lm} model objects.  A set of model objects created by \code{create_model_object}.}
}
\description{
Multiple Regression Assumption Checking
}
\details{
Creates plots and text output to summarize models and check assumptions via objects created by \code{assumptions_check}.  Uses full model with all predictors.
}
\examples{
\dontrun{model_check(freeny_models)}
}

