% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{model_summary_table_binomial}
\alias{model_summary_table_binomial}
\title{Binary Logistic Regression: Summary Table Output}
\usage{
model_summary_table_binomial(models)
}
\arguments{
\item{models}{A list of \code{lm} model objects.  A set of model objects created by \code{create_model_object}.}
}
\description{
Binary Logistic Regression: Summary Table Output
}
\details{
Creates table output to summarize model coefficients for all models in a hierarchical binary logistic regression analysis.
}
\examples{
\dontrun{model_coefficient_table(freeny_models)}
}

