% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{modelCompareMod}
\alias{modelCompareMod}
\title{Modified modelCompare function from lmSupport package.  Modified to suppress
print output.}
\usage{
modelCompareMod(ModelC, ModelA, printOutput = T)
}
\arguments{
\item{ModelC}{A model \code{lm} object.}

\item{ModelA}{A model \code{lm} object.}

\item{printOutput}{Boolean parameter, if \code{TRUE}, print output is
supressed from modelCompare function.}
}
\description{
Modified modelCompare function from lmSupport package.  Modified to suppress
print output.
}
\details{
This is a modification of the modelCompare function that allows
  print output to be suppressed.
}

