% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{model_coefficient_table}
\alias{model_coefficient_table}
\title{Hierarchical regression: Coefficient table output}
\usage{
model_coefficient_table(models)
}
\arguments{
\item{models}{A list of \code{lm} model objects.  A set of model objects
created by \code{create_model_object}.}
}
\description{
Hierarchical regression: Coefficient table output
}
\details{
Creates table output to summarize model coefficients for all models
  in a hierarchical regression analysis.
}
\examples{
freeny_model_formulas <- create_formula_objects("y", c("lag.quarterly.revenue")
, c("price.index"))
freeny_models <- create_model_objects(freeny_model_formulas,
dataset = freeny)
model_coefficient_table(freeny_models)
}

