% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{model_summary_table_binomial}
\alias{model_summary_table_binomial}
\title{Binary Logistic Regression: Model Summary Output}
\usage{
model_summary_table_binomial(models)
}
\arguments{
\item{models}{A list of \code{lm} model objects.  A set of model objects
created by \code{create_model_object}.}
}
\description{
Binary Logistic Regression: Model Summary Output
}
\details{
Creates summary table with pseudo R^2 values for the binary logistic
  regression model objects
}
\examples{
formulas <- create_formula_objects("am", c("hp", "mpg"), c("disp"),
c("drat"))
mtcars_models <- create_model_objects(formulas, data=mtcars,
type="binomial")
last_model <- mtcars_models[[length(mtcars_models)]]
}

